requirements.txt
streamlit
gspread
oauth2client
PyDrive2
google-generativeai
Pillow
